+++ src/main/java/blackjack/domain/player/Participants.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.player.Participant;
 import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
