+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.player.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
